fix: Update decomposition of file structure associated with healthcheck.
