+++ {{cookiecutter.bot_project_name}}/app/api/endpoints/botx.py
@@ -1,7 +1,6 @@
 """Endpoints for communication with botx."""
 
 from http import HTTPStatus
-from typing import Optional
 
 from fastapi import APIRouter, Request
 from fastapi.responses import JSONResponse
@@ -13,37 +12,11 @@ from pybotx import (
     build_command_accepted_response,
 )
 
-from app.api.dependencies.bot import (
-    bot_dependency,
-    check_db_connection_dependency,
-    check_redis_connection_dependency,
-)
+from app.api.dependencies.bot import bot_dependency
 from app.logger import logger
-from app.services.healthcheck import (
-    HealthCheckResponse,
-    HealthCheckResponseBuilder,
-    HealthCheckServiceResult,
-)
 from app.settings import settings
 
 router = APIRouter()
-
-
-@router.get("/healthcheck")
-async def healthcheck(
-    redis_connection_error: Optional[str] = check_redis_connection_dependency,
-    db_connection_error: Optional[str] = check_db_connection_dependency,
-) -> HealthCheckResponse:
-    """Check the health of the bot and services."""
-    healthcheck_builder = HealthCheckResponseBuilder()
-    healthcheck_builder.add_healthcheck_result(
-        HealthCheckServiceResult(name="postgres", error=db_connection_error)
-    )
-    healthcheck_builder.add_healthcheck_result(
-        HealthCheckServiceResult(name="redis", error=redis_connection_error)
-    )
-
-    return healthcheck_builder.build()
 
 
 @router.post("/command")
