+++ {{cookiecutter.bot_project_name}}/app/api/routers.py
@@ -2,7 +2,9 @@
 from fastapi import APIRouter
 
 from app.api.endpoints.botx import router as bot_router
+from app.api.endpoints.healthcheck import router as healthcheck_router
 
 router = APIRouter()
 
-router.include_router(bot_router, tags=["botx"])
+router.include_router(healthcheck_router)
+router.include_router(bot_router)
